test/e2e/framework: move more helpers
