+++ test/e2e/conformance/crd_partial_metadata_test.go
@@ -40,6 +40,7 @@ import (
 	conditionsv1alpha1 "github.com/kcp-dev/kcp/sdk/apis/third_party/conditions/apis/conditions/v1alpha1"
 	"github.com/kcp-dev/kcp/sdk/apis/third_party/conditions/util/conditions"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
+	frameworkhelpers "github.com/kcp-dev/kcp/test/e2e/framework/helpers"
 )
 
 func TestPartialMetadataCRD(t *testing.T) {
@@ -116,7 +117,7 @@ func TestPartialMetadataCRD(t *testing.T) {
 		t.Log("List resources with partial object metadata")
 		metadataClient, err := metadata.NewForConfig(cfg)
 		require.NoError(t, err)
-		framework.Eventually(t, func() (success bool, reason string) {
+		frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 			_, err = metadataClient.Cluster(workspacePath).Resource(resource).List(ctx, metav1.ListOptions{})
 			if err != nil {
 				return false, err.Error()
@@ -217,13 +218,13 @@ func TestPartialMetadataSameCRDMultipleWorkspaces(t *testing.T) {
 	require.NoError(t, err)
 
 	t.Logf("Workspace %s: waiting for ready CRD with version %s", workspace1.Name, crd.Spec.Versions[0].Name)
-	framework.EventuallyCondition(t, func() (conditions.Getter, error) {
+	frameworkhelpers.EventuallyCondition(t, func() (conditions.Getter, error) {
 		crd, err := crdClusterClient.ApiextensionsV1().CustomResourceDefinitions().Cluster(workspace1Path).Get(ctx, crdName, metav1.GetOptions{})
 		if err != nil {
 			return nil, err
 		}
 		return &crdConditionsAdapter{CustomResourceDefinition: crd}, nil
-	}, framework.Is(conditionsv1alpha1.ConditionType(apiextensionsv1.Established)), wait.ForeverTestTimeout, 100*time.Millisecond)
+	}, frameworkhelpers.Is(conditionsv1alpha1.ConditionType(apiextensionsv1.Established)), wait.ForeverTestTimeout, 100*time.Millisecond)
 
 	// Create ws2. Using the root shard because both ws1 and ws2 must be on the same shard to exercise this issue.
 	workspace2Path, workspace2 := framework.NewOrganizationFixture(t, server, framework.WithRootShard())
@@ -236,13 +237,13 @@ func TestPartialMetadataSameCRDMultipleWorkspaces(t *testing.T) {
 	require.NoError(t, err)
 
 	t.Logf("Workspace %s: waiting for ready CRD with version %s", workspace2.Name, crd.Spec.Versions[0].Name)
-	framework.EventuallyCondition(t, func() (conditions.Getter, error) {
+	frameworkhelpers.EventuallyCondition(t, func() (conditions.Getter, error) {
 		crd, err := crdClusterClient.ApiextensionsV1().CustomResourceDefinitions().Cluster(workspace2Path).Get(ctx, crdName, metav1.GetOptions{})
 		if err != nil {
 			return nil, err
 		}
 		return &crdConditionsAdapter{CustomResourceDefinition: crd}, nil
-	}, framework.Is(conditionsv1alpha1.ConditionType(apiextensionsv1.Established)), wait.ForeverTestTimeout, 100*time.Millisecond)
+	}, frameworkhelpers.Is(conditionsv1alpha1.ConditionType(apiextensionsv1.Established)), wait.ForeverTestTimeout, 100*time.Millisecond)
 
 	dynamicClusterClient, err := dynamic.NewForConfig(cfg)
 	require.NoError(t, err)
@@ -304,7 +305,7 @@ func TestPartialMetadataSameCRDMultipleWorkspaces(t *testing.T) {
 	require.NoError(t, err)
 
 	t.Logf("Listing partial metadata for version v1")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		metadataList, err := metadataClusterClient.Cluster(logicalcluster.Wildcard).Resource(gvrV1).List(ctx, metav1.ListOptions{})
 		if err != nil {
 			return false, err.Error()
@@ -325,7 +326,7 @@ func TestPartialMetadataSameCRDMultipleWorkspaces(t *testing.T) {
 
 	// Do a partial metadata list for v2
 	t.Logf("Listing partial metadata for version v2")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		metadataList, err := metadataClusterClient.Cluster(logicalcluster.Wildcard).Resource(gvrV2).List(ctx, metav1.ListOptions{})
 		if err != nil {
 			return false, err.Error()
