+++ test/e2e/reconciler/apiexportendpointslice/apiexportendpointslice_test.go
@@ -43,6 +43,7 @@ import (
 	topologyv1alpha1 "github.com/kcp-dev/kcp/sdk/apis/topology/v1alpha1"
 	kcpclientset "github.com/kcp-dev/kcp/sdk/client/clientset/versioned/cluster"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
+	frameworkhelpers "github.com/kcp-dev/kcp/test/e2e/framework/helpers"
 )
 
 //go:embed *.yaml
@@ -112,7 +113,7 @@ func TestAPIExportEndpointSliceWithPartition(t *testing.T) {
 
 	var sliceName string
 	t.Logf("Retrying to create the APIExportEndpointSlice after the APIExport has been created")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		created, err := sliceClient.Cluster(partitionClusterPath).Create(ctx, slice, metav1.CreateOptions{})
 		if err != nil {
 			return false, err.Error()
@@ -121,7 +122,7 @@ func TestAPIExportEndpointSliceWithPartition(t *testing.T) {
 		return true, ""
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "expected APIExportEndpointSlice creation to succeed")
 
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		slice, err = kcpClusterClient.Cluster(partitionClusterPath).ApisV1alpha1().APIExportEndpointSlices().Get(ctx, sliceName, metav1.GetOptions{})
 		require.NoError(t, err)
 
@@ -142,7 +143,7 @@ func TestAPIExportEndpointSliceWithPartition(t *testing.T) {
 	})
 	require.NoError(t, err, "error updating APIExportEndpointSlice")
 
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		slice, err = kcpClusterClient.Cluster(partitionClusterPath).ApisV1alpha1().APIExportEndpointSlices().Get(ctx, sliceName, metav1.GetOptions{})
 		require.NoError(t, err)
 		if conditions.IsFalse(slice, apisv1alpha1.PartitionValid) && conditions.GetReason(slice, apisv1alpha1.PartitionValid) == apisv1alpha1.PartitionInvalidReferenceReason {
@@ -157,7 +158,7 @@ func TestAPIExportEndpointSliceWithPartition(t *testing.T) {
 	_, err = partitionClient.Cluster(partitionClusterPath).Create(ctx, partition, metav1.CreateOptions{})
 	require.NoError(t, err, "error creating Partition")
 
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		slice, err = kcpClusterClient.Cluster(partitionClusterPath).ApisV1alpha1().APIExportEndpointSlices().Get(ctx, sliceName, metav1.GetOptions{})
 		require.NoError(t, err)
 		if conditions.IsTrue(slice, apisv1alpha1.PartitionValid) {
@@ -255,7 +256,7 @@ func TestAPIBindingEndpointSlicesSharded(t *testing.T) {
 			kcpClusterClient, err := kcpclientset.NewForConfig(cfg)
 			require.NoError(t, err, "failed to construct kcp cluster client for server")
 
-			framework.Eventually(t, func() (bool, string) {
+			frameworkhelpers.Eventually(t, func() (bool, string) {
 				_, err = kcpClusterClient.Cluster(consumerPath).ApisV1alpha1().APIBindings().Create(ctx, apiBinding, metav1.CreateOptions{})
 				return err == nil, fmt.Sprintf("Error creating APIBinding: %v", err)
 			}, wait.ForeverTestTimeout, time.Millisecond*100)
@@ -296,7 +297,7 @@ func TestAPIBindingEndpointSlicesSharded(t *testing.T) {
 		require.NoError(t, err)
 
 		// Partition should be created
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			partitions, err := kcpClusterClient.Cluster(providerPath).TopologyV1alpha1().Partitions().List(ctx, metav1.ListOptions{})
 			if err == nil && len(partitions.Items) == 1 {
 				partition = &partitions.Items[0]
@@ -326,7 +327,7 @@ func TestAPIBindingEndpointSlicesSharded(t *testing.T) {
 		require.NoError(t, err)
 
 		// we should have 1 APIExportEndpointSlice with 1 APIExportEndpoint as we bound only once.
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			slice, err := kcpClusterClient.Cluster(providerPath).ApisV1alpha1().APIExportEndpointSlices().Get(ctx, "shared-cowboys", metav1.GetOptions{})
 			if len(slice.Status.APIExportEndpoints) == 1 {
 				return true, ""
@@ -362,7 +363,7 @@ func TestAPIBindingEndpointSlicesSharded(t *testing.T) {
 			kcpClusterClient, err := kcpclientset.NewForConfig(cfg)
 			require.NoError(t, err, "failed to construct kcp cluster client for server")
 
-			framework.Eventually(t, func() (bool, string) {
+			frameworkhelpers.Eventually(t, func() (bool, string) {
 				_, err = kcpClusterClient.Cluster(consumerPath).ApisV1alpha1().APIBindings().Create(ctx, apiBinding, metav1.CreateOptions{})
 				return err == nil, fmt.Sprintf("Error creating APIBinding: %v", err)
 			}, wait.ForeverTestTimeout, time.Millisecond*500)
@@ -371,7 +372,7 @@ func TestAPIBindingEndpointSlicesSharded(t *testing.T) {
 
 	t.Logf("Check that APIExportEndpointSlices has 2 virtual workspaces")
 	{
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			kcpClusterClient, err := kcpclientset.NewForConfig(cfg)
 			require.NoError(t, err, "failed to construct kcp cluster client for server")
 
@@ -388,7 +389,7 @@ func TestAPIBindingEndpointSlicesSharded(t *testing.T) {
 		kcpClusterClient, err := kcpclientset.NewForConfig(cfg)
 		require.NoError(t, err, "failed to construct kcp cluster client for server")
 
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			err := kcpClusterClient.Cluster(consumerPath).ApisV1alpha1().APIBindings().Delete(ctx, "cowboys", metav1.DeleteOptions{})
 			return err == nil, fmt.Sprintf("Error deleting APIBinding: %v", err)
 		}, wait.ForeverTestTimeout, time.Millisecond*500)
@@ -396,7 +397,7 @@ func TestAPIBindingEndpointSlicesSharded(t *testing.T) {
 
 	t.Logf("Check that APIExportEndpointSlices has 1 virtual workspaces")
 	{
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			kcpClusterClient, err := kcpclientset.NewForConfig(cfg)
 			require.NoError(t, err, "failed to construct kcp cluster client for server")
