+++ test/e2e/reconciler/partitionset/partitionset_test.go
@@ -37,6 +37,7 @@ import (
 	topologyv1alpha1 "github.com/kcp-dev/kcp/sdk/apis/topology/v1alpha1"
 	kcpclientset "github.com/kcp-dev/kcp/sdk/client/clientset/versioned/cluster"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
+	frameworkhelpers "github.com/kcp-dev/kcp/test/e2e/framework/helpers"
 )
 
 func TestPartitionSet(t *testing.T) {
@@ -76,7 +77,7 @@ func TestPartitionSet(t *testing.T) {
 	}
 	partitionSet, err = partitionSetClient.Cluster(partitionClusterPath).Create(ctx, partitionSet, metav1.CreateOptions{})
 	require.NoError(t, err, "error creating partitionSet")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitionSet, err = partitionSetClient.Cluster(partitionClusterPath).Get(ctx, partitionSet.Name, metav1.GetOptions{})
 		require.NoError(t, err, "error retrieving partitionSet")
 		if conditions.IsTrue(partitionSet, topologyv1alpha1.PartitionSetValid) && conditions.IsTrue(partitionSet, topologyv1alpha1.PartitionsReady) {
@@ -112,7 +113,7 @@ func TestPartitionSet(t *testing.T) {
 		err = shardClient.Cluster(core.RootCluster.Path()).Delete(ctx, shard1a.Name, metav1.DeleteOptions{})
 		require.NoError(t, err, "error deleting shard")
 	}()
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitionSet, err = partitionSetClient.Cluster(partitionClusterPath).Get(ctx, partitionSet.Name, metav1.GetOptions{})
 		require.NoError(t, err, "error retrieving partitionSet")
 		if conditions.IsTrue(partitionSet, topologyv1alpha1.PartitionsReady) && partitionSet.Status.Count == uint(1) {
@@ -120,7 +121,7 @@ func TestPartitionSet(t *testing.T) {
 		}
 		return false, fmt.Sprintf("expected 1 partition, but got %d", partitionSet.Status.Count)
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "expected the partition count to be 1")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitions, err = partitionClient.Cluster(partitionClusterPath).List(ctx, metav1.ListOptions{})
 		require.NoError(t, err, "error retrieving partitions")
 		if len(partitions.Items) == 1 {
@@ -152,7 +153,7 @@ func TestPartitionSet(t *testing.T) {
 		err = shardClient.Cluster(core.RootCluster.Path()).Delete(ctx, shard2.Name, metav1.DeleteOptions{})
 		require.NoError(t, err, "error deleting shard")
 	}()
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitionSet, err = partitionSetClient.Cluster(partitionClusterPath).Get(ctx, partitionSet.Name, metav1.GetOptions{})
 		require.NoError(t, err, "error retrieving partitionSet")
 		if conditions.IsTrue(partitionSet, topologyv1alpha1.PartitionsReady) && partitionSet.Status.Count == uint(2) {
@@ -160,7 +161,7 @@ func TestPartitionSet(t *testing.T) {
 		}
 		return false, fmt.Sprintf("expected 2 partitions, but got %d", partitionSet.Status.Count)
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "expected the partitions to be ready and their count to be 2")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitions, err = partitionClient.Cluster(partitionClusterPath).List(ctx, metav1.ListOptions{})
 		require.NoError(t, err, "error retrieving partitions")
 		if len(partitions.Items) == 2 {
@@ -179,7 +180,7 @@ func TestPartitionSet(t *testing.T) {
 	}
 	_, err = shardClient.Cluster(core.RootCluster.Path()).Update(ctx, shard2, metav1.UpdateOptions{})
 	require.NoError(t, err, "error updating shard")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitionSet, err = partitionSetClient.Cluster(partitionClusterPath).Get(ctx, partitionSet.Name, metav1.GetOptions{})
 		require.NoError(t, err, "error retrieving partitionSet")
 		if conditions.IsTrue(partitionSet, topologyv1alpha1.PartitionsReady) && partitionSet.Status.Count == uint(1) {
@@ -187,7 +188,7 @@ func TestPartitionSet(t *testing.T) {
 		}
 		return false, fmt.Sprintf("expected 1 partition, but got %d", partitionSet.Status.Count)
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "expected the partition count to become 1")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitions, err = partitionClient.Cluster(partitionClusterPath).List(ctx, metav1.ListOptions{})
 		require.NoError(t, err, "error retrieving partitions")
 		if len(partitions.Items) == 1 {
@@ -217,7 +218,7 @@ func TestPartitionSet(t *testing.T) {
 		err = shardClient.Cluster(core.RootCluster.Path()).Delete(ctx, shard3.Name, metav1.DeleteOptions{})
 		require.NoError(t, err, "error deleting shard")
 	}()
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitionSet, err = partitionSetClient.Cluster(partitionClusterPath).Get(ctx, partitionSet.Name, metav1.GetOptions{})
 		require.NoError(t, err, "error retrieving partitionSet")
 		if conditions.IsTrue(partitionSet, topologyv1alpha1.PartitionsReady) && partitionSet.Status.Count == uint(2) {
@@ -225,7 +226,7 @@ func TestPartitionSet(t *testing.T) {
 		}
 		return false, fmt.Sprintf("expected 2 partitions, but got %d", partitionSet.Status.Count)
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "expected the partition count to become 2")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitions, err = partitionClient.Cluster(partitionClusterPath).List(ctx, metav1.ListOptions{})
 		require.NoError(t, err, "error retrieving partitions")
 		if len(partitions.Items) == 2 {
@@ -241,7 +242,7 @@ func TestPartitionSet(t *testing.T) {
 	}
 	_, err = shardClient.Cluster(core.RootCluster.Path()).Update(ctx, shard3, metav1.UpdateOptions{})
 	require.NoError(t, err, "error updating shard")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitionSet, err = partitionSetClient.Cluster(partitionClusterPath).Get(ctx, partitionSet.Name, metav1.GetOptions{})
 		require.NoError(t, err, "error retrieving partitionSet")
 		if conditions.IsTrue(partitionSet, topologyv1alpha1.PartitionsReady) && partitionSet.Status.Count == uint(1) {
@@ -249,7 +250,7 @@ func TestPartitionSet(t *testing.T) {
 		}
 		return false, fmt.Sprintf("expected 1 partition, but got %d", partitionSet.Status.Count)
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "expected the partition count to become 1")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitions, err = partitionClient.Cluster(partitionClusterPath).List(ctx, metav1.ListOptions{})
 		require.NoError(t, err, "error retrieving partitions")
 		if len(partitions.Items) == 1 {
@@ -382,7 +383,7 @@ func TestPartitionSetAdmission(t *testing.T) {
 		require.NoError(t, err, "error deleting shard")
 	}()
 	var partitions *topologyv1alpha1.PartitionList
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		partitions, err = partitionClient.Cluster(partitionClusterPath).List(ctx, metav1.ListOptions{})
 		require.NoError(t, err, "error retrieving partitions")
 		if len(partitions.Items) == 1 {
