+++ test/e2e/virtual/apiexport/authorizer_test.go
@@ -57,6 +57,7 @@ import (
 	kcpclientset "github.com/kcp-dev/kcp/sdk/client/clientset/versioned/cluster"
 	"github.com/kcp-dev/kcp/test/e2e/fixtures/wildwest/apis/wildwest"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
+	frameworkhelpers "github.com/kcp-dev/kcp/test/e2e/framework/helpers"
 )
 
 func TestAPIExportAuthorizers(t *testing.T) {
@@ -130,9 +131,9 @@ func TestAPIExportAuthorizers(t *testing.T) {
 	t.Logf("get the sheriffs apiexport's generated identity hash")
 	serviceProvider1AdminClient, err := kcpclientset.NewForConfig(serviceProvider1Admin)
 	require.NoError(t, err)
-	framework.EventuallyCondition(t, func() (conditions.Getter, error) {
+	frameworkhelpers.EventuallyCondition(t, func() (conditions.Getter, error) {
 		return serviceProvider1AdminClient.Cluster(serviceProvider1Path).ApisV1alpha1().APIExports().Get(ctx, "wild.wild.west", metav1.GetOptions{})
-	}, framework.Is(apisv1alpha1.APIExportIdentityValid))
+	}, frameworkhelpers.Is(apisv1alpha1.APIExportIdentityValid))
 
 	sheriffExport, err := serviceProvider1AdminClient.Cluster(serviceProvider1Path).ApisV1alpha1().APIExports().Get(ctx, "wild.wild.west", metav1.GetOptions{})
 	require.NoError(t, err)
@@ -197,7 +198,7 @@ func TestAPIExportAuthorizers(t *testing.T) {
 	))
 
 	t.Logf("bind cowboys and claimed sherriffs in the tenant workspace %q", tenantPath)
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		err := apply(t, ctx, tenantPath, tenantUser,
 			&apisv1alpha1.APIBinding{
 				ObjectMeta: metav1.ObjectMeta{
@@ -252,7 +253,7 @@ func TestAPIExportAuthorizers(t *testing.T) {
 	t.Logf("Make sure [%q, %q] API groups shows up in consumer workspace %q group discovery", wildwest.GroupName, "wild.wild.west", tenantPath)
 	tenantUserWorkspaceKcpClient, err := kcpclientset.NewForConfig(tenantUser)
 	require.NoError(t, err)
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		groups, err := tenantUserWorkspaceKcpClient.Cluster(tenantPath).Discovery().ServerGroups()
 		if err != nil {
 			return false, fmt.Sprintf("error retrieving consumer workspace %q group discovery: %v", tenantPath, err)
@@ -284,7 +285,7 @@ func TestAPIExportAuthorizers(t *testing.T) {
 	t.Logf("Waiting for cowboys CRD to be ready in tenant workspace %q", tenantShadowCRDPath)
 	tenantUserAPIExtensionsClient, err := kcpapiextensionsclientset.NewForConfig(tenantUser)
 	require.NoError(t, err)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		cowboysCRD, err := tenantUserAPIExtensionsClient.Cluster(tenantShadowCRDPath).ApiextensionsV1().CustomResourceDefinitions().Get(ctx, "cowboys.wildwest.dev", metav1.GetOptions{})
 		if err != nil {
 			return false, fmt.Sprintf("error creating API binding: %v", err)
@@ -296,7 +297,7 @@ func TestAPIExportAuthorizers(t *testing.T) {
 	}, wait.ForeverTestTimeout, time.Millisecond*100, "waiting for cowboys CRD to become established failed")
 
 	t.Logf("Create a cowboys APIBinding in consumer workspace %q that points to the today-cowboys export from %q but shadows a local cowboys CRD at the same time", tenantShadowCRDPath, serviceProvider2Path)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		err := apply(t, ctx, tenantShadowCRDPath, tenantUser,
 			&apisv1alpha1.APIBinding{
 				ObjectMeta: metav1.ObjectMeta{
@@ -338,14 +339,14 @@ func TestAPIExportAuthorizers(t *testing.T) {
 	t.Logf("Waiting for cowboys APIBinding in consumer workspace %q to have the condition %q mentioning the conflict with the shadowing local cowboys CRD", tenantShadowCRDPath, apisv1alpha1.BindingUpToDate)
 	tenantUserKcpClient, err := kcpclientset.NewForConfig(tenantUser)
 	require.NoError(t, err)
-	framework.EventuallyCondition(t, func() (conditions.Getter, error) {
+	frameworkhelpers.EventuallyCondition(t, func() (conditions.Getter, error) {
 		return tenantUserKcpClient.Cluster(tenantShadowCRDPath).ApisV1alpha1().APIBindings().Get(ctx, "cowboys", metav1.GetOptions{})
-	}, framework.IsNot(apisv1alpha1.BindingUpToDate).WithReason(apisv1alpha1.NamingConflictsReason))
+	}, frameworkhelpers.IsNot(apisv1alpha1.BindingUpToDate).WithReason(apisv1alpha1.NamingConflictsReason))
 
 	// Have to do this with Eventually because the RBAC for the maximal permission policy can be slow to propagate via
 	// the cache server.
 	t.Logf("Creating cowboy (via APIBinding) in %q", tenantPath)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		err := apply(t, ctx, tenantPath, tenantUser, `
 apiVersion: wildwest.dev/v1alpha1
 kind: Cowboy
@@ -372,7 +373,7 @@ metadata:
 	serviceProvider2AdminApiExportVWCfg := rest.CopyConfig(serviceProvider2Admin)
 	serviceProvider2AdminClient, err := kcpclientset.NewForConfig(serviceProvider2Admin)
 	require.NoError(t, err)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		apiExport, err := serviceProvider2AdminClient.Cluster(serviceProvider2Path).ApisV1alpha1().APIExports().Get(ctx, "today-cowboys", metav1.GetOptions{})
 		require.NoError(t, err)
 		var found bool
@@ -386,7 +387,7 @@ metadata:
 	require.NoError(t, err)
 
 	t.Logf("verify that service-provider-2-admin cannot list sheriffs resources via virtual apiexport apiserver because we have no local maximal permissions yet granted")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		_, err = serviceProvider2DynamicVWClientForTenantWorkspace.Resource(schema.GroupVersionResource{Version: "v1", Resource: "sheriffs", Group: "wild.wild.west"}).List(ctx, metav1.ListOptions{})
 		if err == nil {
 			return false, "expected an error but got nil"
@@ -394,7 +395,7 @@ metadata:
 		return strings.Contains(err.Error(), `sheriffs.wild.wild.west is forbidden: User "service-provider-2-admin" cannot list resource "sheriffs" in API group "wild.wild.west" at the cluster scope: access denied`), fmt.Sprintf("unexpected error: %v", err)
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "service-provider-2-admin must not be allowed to list sheriff resources")
 
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		_, err = serviceProvider2DynamicVWClientForTenantWorkspace.Resource(schema.GroupVersionResource{Version: "v1", Resource: "configmaps"}).List(ctx, metav1.ListOptions{})
 		if err != nil {
 			return false, err.Error()
@@ -421,7 +422,7 @@ metadata:
 		{Version: "v1", Resource: "configmaps"},
 		{Version: "v1alpha1", Resource: "sheriffs", Group: "wild.wild.west"},
 	}
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		for _, gvr := range claimedGVRs {
 			if _, err := serviceProvider2DynamicVWClientForTenantWorkspace.Resource(gvr).List(ctx, metav1.ListOptions{}); err != nil {
 				return false, fmt.Sprintf("error while waiting to list %q: %v", gvr, err)
@@ -431,7 +432,7 @@ metadata:
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "listing claimed resources failed")
 
 	t.Logf("verify that service-provider-2-admin can lists sherriffs resources in the tenant workspace %q via the virtual apiexport apiserver", tenantPath)
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		_, err = serviceProvider2DynamicVWClientForTenantWorkspace.Cluster(logicalcluster.Name(tenantWorkspace.Spec.Cluster).Path()).Resource(schema.GroupVersionResource{Version: "v1alpha1", Resource: "sheriffs", Group: "wild.wild.west"}).List(ctx, metav1.ListOptions{})
 		if err != nil {
 			return false, fmt.Sprintf("error while waiting to list sherriffs: %v", err)
@@ -443,7 +444,7 @@ metadata:
 	shadowVWCfg := rest.CopyConfig(serviceProvider2Admin)
 	shadowVWClient, err := kcpclientset.NewForConfig(serviceProvider2Admin)
 	require.NoError(t, err)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		apiExport, err := shadowVWClient.Cluster(serviceProvider2Path).ApisV1alpha1().APIExports().Get(ctx, "today-cowboys", metav1.GetOptions{})
 		require.NoError(t, err)
 		var found bool
@@ -457,7 +458,7 @@ metadata:
 	require.NoError(t, err)
 
 	t.Logf("verify that service-provider-2-admin cannot list CRD shadowed cowboy resources in the tenant workspace %q via the virtual apiexport apiserver", tenantShadowCRDPath)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		_, err = serviceProvider2DynamicVWClientForShadowTenantWorkspace.Cluster(logicalcluster.Name(tenantShadowCRDWorkspace.Spec.Cluster).Path()).Resource(schema.GroupVersionResource{Version: "v1alpha1", Resource: "cowboys", Group: "wildwest.dev"}).List(ctx, metav1.ListOptions{})
 		if err == nil {
 			return false, "expected error, got none"
@@ -583,9 +584,9 @@ func TestRootAPIExportAuthorizers(t *testing.T) {
 	require.NoError(t, err)
 
 	t.Logf("Get the root tenancy APIExport's identity hash")
-	framework.EventuallyCondition(t, func() (conditions.Getter, error) {
+	frameworkhelpers.EventuallyCondition(t, func() (conditions.Getter, error) {
 		return kcpClient.Cluster(core.RootCluster.Path()).ApisV1alpha1().APIExports().Get(ctx, "tenancy.kcp.io", metav1.GetOptions{})
-	}, framework.Is(apisv1alpha1.APIExportIdentityValid))
+	}, frameworkhelpers.Is(apisv1alpha1.APIExportIdentityValid))
 
 	tenancyAPIExport, err := kcpClient.Cluster(core.RootCluster.Path()).ApisV1alpha1().APIExports().Get(ctx, "tenancy.kcp.io", metav1.GetOptions{})
 	require.NoError(t, err)
@@ -647,15 +648,15 @@ func TestRootAPIExportAuthorizers(t *testing.T) {
 			},
 		},
 	}
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		_, err := userKcpClient.Cluster(userPath).ApisV1alpha1().APIBindings().Create(ctx, apiBinding, metav1.CreateOptions{})
 		return err == nil, fmt.Sprintf("Error creating APIBinding: %v", err)
 	}, wait.ForeverTestTimeout, time.Millisecond*100, "api binding creation failed")
 
 	t.Logf("Wait for the binding to be ready")
-	framework.EventuallyCondition(t, func() (conditions.Getter, error) {
+	frameworkhelpers.EventuallyCondition(t, func() (conditions.Getter, error) {
 		return userKcpClient.Cluster(userPath).ApisV1alpha1().APIBindings().Get(ctx, apiBinding.Name, metav1.GetOptions{})
-	}, framework.Is(apisv1alpha1.InitialBindingCompleted))
+	}, frameworkhelpers.Is(apisv1alpha1.InitialBindingCompleted))
 
 	t.Logf("Get virtual workspace client for service APIExport in workspace %q", servicePath)
 	serviceAPIExportVWCfg := framework.StaticTokenUserConfig(providerUser, rest.CopyConfig(cfg))
@@ -680,7 +681,7 @@ func TestRootAPIExportAuthorizers(t *testing.T) {
 		},
 	}
 
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		_, err = serviceDynamicVWClient.Cluster(userClusterName.Path()).
 			Resource(tenancyv1alpha1.SchemeGroupVersion.WithResource("workspacetypes")).
 			Create(ctx, workspaceType, metav1.CreateOptions{})
@@ -699,7 +700,7 @@ func vwURL(t *testing.T, kcpClusterClient kcpclientset.ClusterInterface, path lo
 	t.Cleanup(cancelFunc)
 
 	var vwURL string
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		export, err := kcpClusterClient.Cluster(path).ApisV1alpha1().APIExports().Get(ctx, export, metav1.GetOptions{})
 		require.NoError(t, err)
 		urls := framework.ExportVirtualWorkspaceURLs(export)
