+++ test/e2e/virtual/apiexport/binding_test.go
@@ -33,6 +33,7 @@ import (
 
 	kcpclientset "github.com/kcp-dev/kcp/sdk/client/clientset/versioned/cluster"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
+	frameworkhelpers "github.com/kcp-dev/kcp/test/e2e/framework/helpers"
 )
 
 func TestBinding(t *testing.T) {
@@ -84,7 +85,7 @@ spec: {}
 `))
 
 	t.Logf("Binding to 'api-manager' APIExport succeeds because service-provider user is admin in 'service-provider' workspace")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		err = apply(t, ctx, consumerWorkspacePath, serviceProviderUser, fmt.Sprintf(`
 apiVersion: apis.kcp.io/v1alpha1
 kind: APIBinding
@@ -108,7 +109,7 @@ spec:
 	}, wait.ForeverTestTimeout, 1000*time.Millisecond, "waiting on binding 'api-manager' export")
 
 	t.Logf("Binding directly to 'restricted-service' APIExport should be forbidden")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		err = apply(t, ctx, consumerWorkspacePath, serviceProviderUser, fmt.Sprintf(`
 apiVersion: apis.kcp.io/v1alpha1
 kind: APIBinding
@@ -129,7 +130,7 @@ spec:
 
 	t.Logf("Waiting for 'api-manager' APIExport virtual workspace URL")
 	serviceProviderVirtualWorkspaceConfig := rest.CopyConfig(serviceProviderUser)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		apiExport, err := kcpClient.Cluster(serviceWorkspacePath).ApisV1alpha1().APIExports().Get(ctx, "api-manager", metav1.GetOptions{})
 		if err != nil {
 			return false, fmt.Sprintf("waiting on apiexport to be available %v", err.Error())
@@ -142,7 +143,7 @@ spec:
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "waiting on virtual workspace to be ready")
 
 	t.Logf("Binding to 'restricted-service' APIExport through 'api-manager' APIExport virtual workspace is forbidden")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		err = apply(t, ctx, logicalcluster.Name(consumerWorkspace.Spec.Cluster).Path(), serviceProviderVirtualWorkspaceConfig, fmt.Sprintf(`
 apiVersion: apis.kcp.io/v1alpha1
 kind: APIBinding
@@ -191,7 +192,7 @@ roleRef:
 `))
 
 	t.Logf("Binding to 'restricted-service' APIExport through 'api-manager' APIExport virtual workspace succeeds, proving that the service provider identity is used through the APIExport virtual workspace")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		err := apply(t, ctx, logicalcluster.Name(consumerWorkspace.Spec.Cluster).Path(), serviceProviderVirtualWorkspaceConfig, fmt.Sprintf(`
 apiVersion: apis.kcp.io/v1alpha1
 kind: APIBinding
