+++ test/e2e/virtual/apiexport/virtualworkspace_test.go
@@ -67,6 +67,7 @@ import (
 	wildwestv1alpha1ac "github.com/kcp-dev/kcp/test/e2e/fixtures/wildwest/client/applyconfiguration/wildwest/v1alpha1"
 	wildwestclientset "github.com/kcp-dev/kcp/test/e2e/fixtures/wildwest/client/clientset/versioned/cluster"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
+	frameworkhelpers "github.com/kcp-dev/kcp/test/e2e/framework/helpers"
 )
 
 func TestAPIExportVirtualWorkspace(t *testing.T) {
@@ -105,7 +106,7 @@ func TestAPIExportVirtualWorkspace(t *testing.T) {
 
 	t.Logf("Waiting for APIExport to have a virtual workspace URL for the bound workspace %q", consumerWorkspace.Name)
 	apiExportVWCfg := rest.CopyConfig(cfg)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		apiExport, err := kcpClients.Cluster(serviceProviderPath).ApisV1alpha1().APIExports().Get(ctx, "today-cowboys", metav1.GetOptions{})
 		require.NoError(t, err)
 		var found bool
@@ -148,7 +149,7 @@ func TestAPIExportVirtualWorkspace(t *testing.T) {
 		admit(t, kubeClusterClient.Cluster(serviceProviderPath), "user-1-vw-list-get", "user-1", "User", []string{"list", "get"}, apisv1alpha1.SchemeGroupVersion.Group, "apiexports/content", "")
 
 		t.Logf("Verify that user-1 can now wildcard list cowboys")
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			cbs, err := wwUser1VC.WildwestV1alpha1().Cowboys().List(ctx, metav1.ListOptions{})
 			if apierrors.IsForbidden(err) {
 				return false, fmt.Sprintf("waiting until rbac cache is primed: %v", err)
@@ -191,7 +192,7 @@ func TestAPIExportVirtualWorkspace(t *testing.T) {
 		admit(t, kubeClusterClient.Cluster(serviceProviderPath), "user-1-vw-update", "user-1", "User", []string{"update"}, apisv1alpha1.SchemeGroupVersion.Group, "apiexports/content", "")
 
 		t.Logf("Verify that user-1 can now update cowboys")
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
 				cowboy, err = wwUser1VC.Cluster(consumerClusterName.Path()).WildwestV1alpha1().Cowboys(cowboy.Namespace).Get(ctx, cowboy.Name, metav1.GetOptions{})
 				require.NoError(t, err)
@@ -207,7 +208,7 @@ func TestAPIExportVirtualWorkspace(t *testing.T) {
 		}, wait.ForeverTestTimeout, time.Millisecond*100, "expected user-1 to update cowboys")
 
 		t.Logf("Verify that user-1 can now update status of cowboys")
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
 				cowboy, err = wwUser1VC.Cluster(consumerClusterName.Path()).WildwestV1alpha1().Cowboys(cowboy.Namespace).Get(ctx, cowboy.Name, metav1.GetOptions{})
 				require.NoError(t, err)
@@ -233,7 +234,7 @@ func TestAPIExportVirtualWorkspace(t *testing.T) {
 		admit(t, kubeClusterClient.Cluster(serviceProviderPath), "user-1-vw-create", "user-1", "User", []string{"create"}, apisv1alpha1.SchemeGroupVersion.Group, "apiexports/content", "")
 
 		t.Logf("Verify that user-1 can now create cowboys")
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			_, err = wwUser1VC.Cluster(consumerClusterName.Path()).WildwestV1alpha1().Cowboys("default").Create(ctx, newCowboy("default", "another"), metav1.CreateOptions{})
 			if apierrors.IsForbidden(err) {
 				return false, fmt.Sprintf("waiting until rbac cache is primed: %v", err)
@@ -257,7 +258,7 @@ func TestAPIExportVirtualWorkspace(t *testing.T) {
 		admit(t, kubeClusterClient.Cluster(serviceProviderPath), "user-1-vw-patch", "user-1", "User", []string{"patch"}, apisv1alpha1.SchemeGroupVersion.Group, "apiexports/content", "")
 
 		t.Logf("Verify that user-1 can now patch (application/merge-patch+json) cowboys")
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			patch := `{"spec":{"intent":"3"}}`
 			_, err = wwUser1VC.Cluster(consumerClusterName.Path()).WildwestV1alpha1().Cowboys(cowboy.Namespace).Patch(ctx, cowboy.Name, types.MergePatchType, []byte(patch), metav1.PatchOptions{})
 			if apierrors.IsForbidden(err) {
@@ -295,7 +296,7 @@ func TestAPIExportVirtualWorkspace(t *testing.T) {
 		admit(t, kubeClusterClient.Cluster(serviceProviderPath), "user-1-vw-delete", "user-1", "User", []string{"delete", "deletecollection"}, apisv1alpha1.SchemeGroupVersion.Group, "apiexports/content", "")
 
 		t.Logf("Verify that user-1 can now delete cowboys")
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			err = wwUser1VC.Cluster(consumerClusterName.Path()).WildwestV1alpha1().Cowboys("default").Delete(ctx, "another", metav1.DeleteOptions{})
 			if apierrors.IsForbidden(err) {
 				return false, fmt.Sprintf("waiting until rbac cache is primed: %v", err)
@@ -393,7 +394,7 @@ func TestAPIExportAPIBindingsAccess(t *testing.T) {
 		apiExportKCPClient, err := kcpclientset.NewForConfig(apiExportCfg)
 		require.NoError(t, err, "error creating wildcard kcp client for %s|%s", clusterName, exportName)
 
-		framework.Eventually(t, func() (bool, string) {
+		frameworkhelpers.Eventually(t, func() (bool, string) {
 			bindings, err := apiExportKCPClient.ApisV1alpha1().APIBindings().List(ctx, metav1.ListOptions{})
 			require.NoError(t, err, "error listing bindings for %s|%s", clusterName, exportName)
 
@@ -466,7 +467,7 @@ func TestAPIExportAPIBindingsAccess(t *testing.T) {
 	})
 	patchBytes, err := jsonpatch.CreateMergePatch([]byte(oldData), []byte(newData))
 	require.NoError(t, err, "error creating patch")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		_, err = kcpClusterClient.Cluster(ws1Path).ApisV1alpha1().APIBindings().Patch(ctx, binding1.Name, types.MergePatchType, patchBytes, metav1.PatchOptions{})
 		if err == nil {
 			return true, ""
@@ -523,7 +524,7 @@ func TestAPIExportAPIBindingsAccess(t *testing.T) {
 	newData = toJSON(t, binding1)
 	patchBytes, err = jsonpatch.CreateMergePatch([]byte(oldData), []byte(newData))
 	require.NoError(t, err, "error creating patch")
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		_, err = kcpClusterClient.Cluster(ws1Path).ApisV1alpha1().APIBindings().Patch(ctx, binding1.Name, types.MergePatchType, patchBytes, metav1.PatchOptions{})
 		if err == nil {
 			return true, ""
@@ -580,9 +581,9 @@ func TestAPIExportPermissionClaims(t *testing.T) {
 	apifixtures.CreateSheriffsSchemaAndExport(ctx, t, serviceProviderSheriffsNotUsed, kcpClusterClient, "wild.wild.west", "use the giant spider")
 
 	t.Logf("get the sheriffs apiexport's generated identity hash")
-	framework.EventuallyCondition(t, func() (conditions.Getter, error) {
+	frameworkhelpers.EventuallyCondition(t, func() (conditions.Getter, error) {
 		return kcpClusterClient.Cluster(sheriffProviderPath).ApisV1alpha1().APIExports().Get(ctx, "wild.wild.west", metav1.GetOptions{})
-	}, framework.Is(apisv1alpha1.APIExportIdentityValid))
+	}, frameworkhelpers.Is(apisv1alpha1.APIExportIdentityValid))
 
 	sheriffExport, err := kcpClusterClient.Cluster(sheriffProviderPath).ApisV1alpha1().APIExports().Get(ctx, "wild.wild.west", metav1.GetOptions{})
 	require.NoError(t, err)
@@ -606,7 +607,7 @@ func TestAPIExportPermissionClaims(t *testing.T) {
 
 	t.Logf("Waiting for cowboy APIExport to have a virtual workspace URL for the bound workspaces %q", consumer1Path)
 	consumer1VWCfg := rest.CopyConfig(cfg)
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		apiExport, err := kcpClusterClient.Cluster(claimerPath).ApisV1alpha1().APIExports().Get(ctx, "today-cowboys", metav1.GetOptions{})
 		require.NoError(t, err)
 		var found bool
@@ -682,7 +683,7 @@ func TestAPIExportPermissionClaims(t *testing.T) {
 	require.NoError(t, err, "error patching apibinding to accept all claims")
 
 	t.Logf("Waiting for apibinding status to show applied claims for the sheriff's identity")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		// Get the binding and make sure that observed permission claims are all set.
 		binding, err := kcpClusterClient.Cluster(consumer1Path).ApisV1alpha1().APIBindings().Get(ctx, "cowboys", metav1.GetOptions{})
 		require.NoError(t, err)
@@ -705,7 +706,7 @@ func TestAPIExportPermissionClaims(t *testing.T) {
 	apifixtures.CreateSheriff(ctx, t, dynamicClusterClient, consumer1Path, "wild.wild.west", "in-vw")
 
 	t.Logf("Verify that two sherrifs are eventually returned")
-	framework.Eventually(t, func() (done bool, str string) {
+	frameworkhelpers.Eventually(t, func() (done bool, str string) {
 		ul, err := dynamicVWClusterClient.Resource(sheriffsGVR).List(ctx, metav1.ListOptions{})
 		if err != nil {
 			return false, err.Error()
@@ -750,7 +751,7 @@ func TestAPIExportPermissionClaims(t *testing.T) {
 	require.NoError(t, err)
 
 	t.Logf("Making sure list configmaps is an error")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		_, err := dynamicVWClusterClient.Resource(schema.GroupVersionResource{Group: "", Version: "v1", Resource: "configmaps"}).List(ctx, metav1.ListOptions{})
 		return apierrors.IsNotFound(err), fmt.Sprintf("waiting for an IsNotFound error, but got: %v", err)
 	}, wait.ForeverTestTimeout, 100*time.Millisecond, "didn't get list configmaps error")
@@ -780,7 +781,7 @@ func TestAPIExportPermissionClaims(t *testing.T) {
 	require.NoError(t, err, "error patching apibinding")
 
 	t.Logf("Verify that the list of sheriffs becomes empty")
-	framework.Eventually(t, func() (success bool, reason string) {
+	frameworkhelpers.Eventually(t, func() (success bool, reason string) {
 		list, err := dynamicVWClusterClient.Resource(sheriffsGVR).List(ctx, metav1.ListOptions{})
 		require.NoError(t, err)
 		return len(list.Items) == 0, fmt.Sprintf("waiting for empty list, got: %#v", list.Items)
@@ -1030,7 +1031,7 @@ func bindConsumerToProvider(ctx context.Context, t *testing.T, consumerWorkspace
 	consumerWsClient, err := kcpclientset.NewForConfig(cfg)
 	require.NoError(t, err)
 
-	framework.Eventually(t, func() (bool, string) {
+	frameworkhelpers.Eventually(t, func() (bool, string) {
 		_, err = kcpClients.Cluster(consumerWorkspace).ApisV1alpha1().APIBindings().Create(ctx, apiBinding, metav1.CreateOptions{})
 		return err == nil, fmt.Sprintf("Error creating APIBinding: %v", err)
 	}, wait.ForeverTestTimeout, time.Millisecond*100)
